dio-iot-transfer-learning/requirements.txt
tensorflow>=2.14.0
numpy>=1.24.0
pillow>=10.0.0
paho-mqtt>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
